arranged  database functions by their models
