+++ server/DataBaseFuncs/ToolFunctions.js
@@ -1,0 +1,12 @@
+
+const fs =require('fs');
+
+const DeleteImgFromStorage =(path)=>{
+    fs.unlink(path,(err)=> {
+      console.log(err);
+    })
+  }
+
+  module.exports = {
+    DeleteImgFromStorage
+  }
